Update OIDC backend

+++ manage/update.sh
@@ -2,11 +2,27 @@
 
 # Updates an existing setup.
 
+LOCAL_DIR=".local/simple"
+CONFIG_FILE="$LOCAL_DIR/config.env"
+DB_FILE="$LOCAL_DIR/db.sqlite3"
 CMD="manage/cmd.sh"
 MANAGE="manage/manage.sh"
 
 set -eu # Exit on error and undefined var is error
 
+# Check if config file exist
+if [[ ! -e $CONFIG_FILE ]]; then
+    echo "App config not found: $SETTINGS_FILE" 1>&2
+    exit -1
+fi
+
+# Create DB file so Docker doesn't make it a directory
+if [[ ! -e $DB_FILE ]]; then
+    echo "Creating DB file ..."
+    touch $DB_FILE
+fi
+
+echo
 echo "Updating requirements ..."
 $CMD pip3 install -r requirements/development.txt | egrep -v "^Requirement already satisfied" || true
